requirements.txt
python == 3.11.6
aiogram == 2.19
python-dotenv == 1.0.1
aiohttp == 3.8.6